requirements.txt
azure-identity==1.16.0
azure-mgmt-cognitiveservices==3.0.0
cachetools==5.3.3
chromadb==0.5.0
datasets==2.19.2
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
pandas==2.2.2
pronto==2.5.7
pysqlite3==0.5.2
python-dotenv==1.0.1
rapidfuzz==3.9.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
tabulate==0.9.0
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
transformers[torch]
urllib3==2.2.1
